More beating at phantoms
I really can't comprehend the program at this point, so mostly I'm just
reduced to poking at things to see if adding or subtracting a 1 works.
It's sad that I'm the only one who wants to do this sort of thing but oh
well. If I concentrate for about an hour I can usually get an idea of
how what I'm doing is working, what the numbers are supposed to be etc,
but before long it's gone just like the knowledge Bones acquired in
Spock's Brain.