All examples work
Finally. No structural changes should be needed now that encryption is
working. Will probably roll my own pynacl for direct access to the C API
since it'd be faster than catching an exception for failed decrypts, and
my CTR method doesn't need the nonce in every packet. But otherwise, now
it's just what applications to build on top of it! Need a file format
for what an "email" should be... probably XML.