"Markup" post format
The idea is that the post links to attachments by content hash key
(i.e. root hash + depth) with metadata about the content as attributes
in the link tag. One link tag for all kinds of links, since getting "a
post plus attachments" is a common operation, so iterating through
"link" is better than through "video" and "audio" and "object" and
"embed" and etc.

serve.py doesn't work because http.server sucks and shouldn't be used.